Split html and plot into separate files
